requirements.txt
pyrtlsdr==0.3.0
numpy>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
